test: 변경된 api에 맞게 테스트 코드 수정
